renamed some routes and refactor some code from bottom navigatinion
